Comments +security measures: only creator and admin can edit and delete comments

+++ src/main/java/com/sda/inTeams/controller/CommentController.java
@@ -5,9 +5,11 @@ import com.sda.inTeams.model.Comment.Comment;
 import com.sda.inTeams.model.User.User;
 import com.sda.inTeams.repository.TaskRepository;
 import com.sda.inTeams.repository.UserRepository;
+import com.sda.inTeams.service.AuthorizationService;
 import com.sda.inTeams.service.CommentService;
+import com.sda.inTeams.service.TaskService;
+import com.sda.inTeams.service.UserService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -20,37 +22,22 @@ import java.security.Principal;
 public class CommentController {
 
     private final CommentService commentService;
-    private final TaskRepository taskRepository;
-    private final UserRepository userRepository;
-
-    @GetMapping()
-    public String getTaskComments(Model model, @RequestParam(name = "taskId") long taskId) {
-        try {
-            model.addAttribute("commentList", commentService.getAllByTask(taskId));
-        } catch (InvalidOperation invalidOperation) {
-            invalidOperation.printStackTrace();
-            model.addAttribute("commentList", commentService.getAll());
-        }
-        return "comment-list";
-    }
+    private final TaskService taskService;
+    private final UserService userService;
+    private final AuthorizationService authorizationService;
 
     @GetMapping("/add")
-    public String addCommentForm(Model model, long taskId, long userId, Principal principal) {
+    public String addCommentForm(Model model, Principal principal, long taskId, long userId) {
         try {
-            if (principal instanceof UsernamePasswordAuthenticationToken) {
-                UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken = (UsernamePasswordAuthenticationToken) principal;
-                if (usernamePasswordAuthenticationToken.getPrincipal() instanceof User) {
-                    User user = (User) usernamePasswordAuthenticationToken.getPrincipal();
-                    taskRepository.findById(taskId).orElseThrow(() -> new InvalidOperation("Task not found!"));
-                    userRepository.findById(userId).orElseThrow(() -> new InvalidOperation("User not found!"));
-                    model.addAttribute("newComment", new Comment());
-                    model.addAttribute("ownerId", taskId);
-                    model.addAttribute("creatorId", user.getId());
-                    return "comment-add-form";
-                }
-            }
-        } catch (InvalidOperation invalidOperation) {
-            invalidOperation.printStackTrace();
+            User user = authorizationService.getUserCredentials(principal).orElseThrow();
+            taskService.getByIdOrThrow(taskId);
+            userService.getByIdOrThrow(userId);
+            model.addAttribute("newComment", new Comment());
+            model.addAttribute("ownerId", taskId);
+            model.addAttribute("creatorId", user.getId());
+            return "comment-add-form";
+        } catch (InvalidOperation operation) {
+            operation.printStackTrace();
         }
         return "redirect:/";
     }
@@ -58,40 +45,48 @@ public class CommentController {
     @PostMapping("/add")
     public String addComment(Comment comment, long taskId, long userId) {
         try {
-            comment.setCreator(userRepository.findById(userId).orElseThrow(() -> new InvalidOperation("User not found!")));
-            comment.setTask(taskRepository.findById(taskId).orElseThrow(() -> new InvalidOperation("Task not found!")));
+            comment.setCreator(userService.getByIdOrThrow(userId));
+            comment.setTask(taskService.getByIdOrThrow(taskId));
             commentService.add(comment);
             return "redirect:/task/" + taskId;
         } catch (InvalidOperation invalidOperation) {
             invalidOperation.printStackTrace();
-            return "redirect:/comment/all";
+            return "redirect:/";
         }
     }
 
     @GetMapping("/edit/{id}")
-    public String editComment(Model model, @PathVariable(name = "id") long commentId) {
+    public String editComment(Model model, Principal principal, @PathVariable(name = "id") long commentId) {
         try {
             Comment commentToEdit = commentService.getByIdOrThrow(commentId);
-            model.addAttribute("newComment", commentToEdit);
-            model.addAttribute("ownerId", commentToEdit.getTask().getId());
-            model.addAttribute("creatorId", commentToEdit.getCreator().getId());
-            return "comment-add-form";
+            if (authorizationService.isUserEligibleToEditComment(principal, commentToEdit)) {
+                model.addAttribute("newComment", commentToEdit);
+                model.addAttribute("ownerId", commentToEdit.getTask().getId());
+                model.addAttribute("creatorId", commentToEdit.getCreator().getId());
+                return "comment-add-form";
+            } else {
+                //unauthorized access
+            }
         } catch (InvalidOperation invalidOperation) {
             invalidOperation.printStackTrace();
-            return "redirect:/comment/all";
         }
+        return "redirect:/";
     }
 
     @GetMapping("/delete/{id}")
-    public String deleteComment(@PathVariable(name = "id") long commentId) {
+    public String deleteComment(Principal principal, @PathVariable(name = "id") long commentId) {
         try {
             Comment comment = commentService.getByIdOrThrow(commentId);
-            long taskId = comment.getTask().getId();
-            commentService.delete(commentId);
-            return "redirect:/task/" + taskId;
+            if (authorizationService.isUserEligibleToDeleteComment(principal, comment)) {
+                long taskId = comment.getTask().getId();
+                commentService.delete(commentId);
+                return "redirect:/task/" + taskId;
+            } else {
+                //unauthorized access
+            }
         } catch (InvalidOperation invalidOperation) {
             invalidOperation.printStackTrace();
-            return "redirect:/task/all";
         }
+        return "redirect:/";
     }
 }
